Image-caption-generator-main/requirements.txt
Flask
Pillow
google-generativeai
